Store processing results

+++ src/data/base.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from typing import List
 
 import pandas as pd
+import torch
 
 from src.data.preprocessing import Pipeline
 
@@ -24,6 +25,12 @@ class DatasetLoader(ABC):
         return path
 
     @property
+    def preprocess_directory(self):
+        path = Path.home() / ".ltr_datasets" / "processed"
+        path.mkdir(parents=True, exist_ok=True)
+        return path
+
+    @property
     def dataset_directory(self):
         path = Path.home() / ".ltr_datasets" / "dataset"
         path.mkdir(parents=True, exist_ok=True)
@@ -41,14 +48,24 @@ class DatasetLoader(ABC):
 
     def load(self, split: str) -> pd.DataFrame:
         assert split in self.splits, f"Split must one of {self.splits}"
-        path = self.cache_directory / f"{self.name}-{self.fold}-{split}.parquet"
+        cache_path = self.preprocess_directory / f"{self.name}-{self.fold}-{split}.pt"
 
-        if not path.exists():
-            df = self._parse(split)
-            df.to_parquet(path)
+        if not cache_path.exists():
+            path = self.cache_directory / f"{self.name}-{self.fold}-{split}.parquet"
 
-        df = pd.read_parquet(path)
-        return self.pipeline(df)
+            if not path.exists():
+                df = self._parse(split)
+                df.to_parquet(path)
+
+            df = pd.read_parquet(path)
+            dataset = self.pipeline(df)
+
+            torch.save(dataset, cache_path)
+
+        return torch.load(cache_path)
+
+
+
 
     @property
     @abstractmethod
